+++ backend/controller/cronjobs/cronjobs_integration_test.go
@@ -4,11 +4,14 @@ package cronjobs
 
 import (
 	"context"
+	"os"
+	"path/filepath"
 	"testing"
 	"time"
 
 	db "github.com/TBD54566975/ftl/backend/controller/dal"
 	"github.com/TBD54566975/ftl/backend/controller/sql/sqltest"
+	in "github.com/TBD54566975/ftl/integration"
 	"github.com/TBD54566975/ftl/internal/log"
 	"github.com/alecthomas/assert/v2"
 	"github.com/benbjohnson/clock"
@@ -34,4 +37,24 @@ func TestServiceWithRealDal(t *testing.T) {
 	}
 
 	testServiceWithDal(ctx, t, dal, clk)
+}
+
+func TestCron(t *testing.T) {
+	dir := t.TempDir()
+	// Due to some MacOS magic, /tmp differs between this test code and the
+	// executing module, so we need to pass the file path as an environment
+	// variable.
+	tmpFile := filepath.Join(dir, "cron.txt")
+	t.Setenv("DEST_FILE", tmpFile)
+
+	t.Cleanup(func() { _ = os.Remove(tmpFile) })
+
+	in.Run(t, "",
+		in.CopyModule("cron"),
+		in.Deploy("cron"),
+		func(t testing.TB, ic in.TestContext) {
+			_, err := os.Stat(tmpFile)
+			assert.NoError(t, err)
+		},
+	)
 }
