+++ backend/controller/cronjobs/testdata/go/cron/go.mod
@@ -40,4 +40,4 @@ require (
 	google.golang.org/protobuf v1.34.1 // indirect
 )
 
-replace github.com/TBD54566975/ftl => ../../../..
+replace github.com/TBD54566975/ftl => ../../../../../..
